Renames BeanFactoryProcessor to ObjectFactory and moves it into context package
