+++ src/main/java/org/blyznytsia/context/AnnotationApplicationContext.java
@@ -12,7 +12,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.blyznytsia.exception.NoSuchBeanException;
 import org.blyznytsia.exception.NoUniqueBeanException;
 import org.blyznytsia.model.BeanDefinition;
-import org.blyznytsia.processor.BeanFactoryProcessor;
 import org.blyznytsia.scanner.BeanScanner;
 import org.blyznytsia.validator.BeanValidator;
 import org.reflections.Reflections;
@@ -56,7 +55,7 @@ public class AnnotationApplicationContext implements ApplicationContext {
    * <b>Steps:</b>
    * 1. Initialize {@link Reflections}
    * 2. Initialize and run all implementations of {@link BeanScanner}
-   * 3. Initialize {@link BeanFactoryProcessor} that creates beans
+   * 3. Initialize {@link ObjectFactory} that creates beans
    * and put them into {@link AnnotationApplicationContext#container}
    * </pre>
    *
@@ -66,7 +65,7 @@ public class AnnotationApplicationContext implements ApplicationContext {
     this.reflections = new Reflections(packageName);
     var beanDefinitions = initAndRunScanners(packageName);
     validate(beanDefinitions);
-    new BeanFactoryProcessor(this).initiateContext(beanDefinitions);
+    new ObjectFactory(this).initiateContext(beanDefinitions);
   }
 
   /**
